requirements.txt
python-decouple
flask
flask-cors
redis[hiredis]